code seems to be working. let's add logging.

+++ LoadBalancer.h
@@ -3,13 +3,14 @@
 
 #include <queue>
 #include <vector>
-#include <Request.h>
+#include "Request.h"
+#include "Webserver.h"
 using namespace std;
 
 struct LoadBalancer {
 
     queue<Request*> request_queue;
-    int check_countdown, active_server_count; 
+    size_t check_countdown, active_server_count; 
     vector<Webserver> webservers;
 
     LoadBalancer(int num_webservers, char web_server_type);
