+++ Switch.cpp
@@ -1,4 +1,6 @@
 #include "Switch.h"
+#include "Request.h"
+#include "LoadBalancer.h"
 #include <vector>
 using namespace std;
 
@@ -9,7 +11,7 @@ Switch::Switch(vector<char> types, int servers_per_loadbalancer) {
 }
 
 void Switch::queueRequest(Request* request) {
-    for(int i = 0; i < loadBalancerTypes.size(); i++) {
+    for(size_t i = 0; i < loadBalancerTypes.size(); i++) {
         if(request->type == loadBalancerTypes[i]) {
             loadBalancers[i].addRequest(request);
         }
